[Feat/#141] 리뷰 반영
